+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.0 (August 31, 2018)
+
+* Add disable logging of first arming
+* Add option to send messages with emoticons
+* Update to Seq Apps 4.2.470 and Serilog 2.7.1
+
 ## 1.1.0 (February 2, 2017)
 
 * New switches are logged as: "New switch" to allow differentiation.